Add `get_preset` for westend + polkadot-node-service deps clean up
