+++ polkadot/node/service/Cargo.toml
@@ -18,15 +18,13 @@ sc-consensus-beefy = { workspace = true, default-features = true }
 sc-consensus-grandpa = { workspace = true, default-features = true }
 mmr-gadget = { workspace = true, default-features = true }
 sp-mmr-primitives = { workspace = true, default-features = true }
-sc-block-builder = { workspace = true, default-features = true }
+sp-genesis-builder = { workspace = true, default-features = true }
 sc-chain-spec = { workspace = true, default-features = true }
 sc-client-api = { workspace = true, default-features = true }
-sc-client-db = { workspace = true, default-features = true }
 sc-consensus = { workspace = true, default-features = true }
 sc-consensus-slots = { workspace = true, default-features = true }
 sc-executor = { workspace = true, default-features = true }
 sc-network = { workspace = true, default-features = true }
-sc-network-common = { workspace = true, default-features = true }
 sc-network-sync = { workspace = true, default-features = true }
 sc-transaction-pool = { workspace = true, default-features = true }
 sc-transaction-pool-api = { workspace = true, default-features = true }
@@ -50,22 +48,17 @@ sp-block-builder = { workspace = true, default-features = true }
 sp-blockchain = { workspace = true, default-features = true }
 sp-core = { workspace = true, default-features = true }
 sp-io = { workspace = true, default-features = true }
-sp-keystore = { workspace = true, default-features = true }
 sp-offchain = { workspace = true, default-features = true }
 sp-runtime = { workspace = true, default-features = true }
 sp-session = { workspace = true, default-features = true }
-sp-storage = { workspace = true, default-features = true }
 sp-transaction-pool = { workspace = true, default-features = true }
 pallet-transaction-payment = { workspace = true, default-features = true }
 sp-timestamp = { workspace = true, default-features = true }
 sp-consensus-babe = { workspace = true, default-features = true }
-sp-state-machine = { workspace = true, default-features = true }
 sp-weights = { workspace = true, default-features = true }
 sp-version = { workspace = true, default-features = true }
 
 # Substrate Pallets
-pallet-babe = { workspace = true, default-features = true }
-pallet-staking = { workspace = true, default-features = true }
 pallet-transaction-payment-rpc-runtime-api = { workspace = true, default-features = true }
 frame-metadata-hash-extension = { optional = true, workspace = true, default-features = true }
 frame-system = { workspace = true, default-features = true }
@@ -73,18 +66,15 @@ frame-system = { workspace = true, default-features = true }
 # Substrate Other
 frame-system-rpc-runtime-api = { workspace = true, default-features = true }
 prometheus-endpoint = { workspace = true, default-features = true }
-frame-support = { workspace = true, default-features = true }
 frame-benchmarking-cli = { workspace = true, default-features = true }
 frame-benchmarking = { workspace = true, default-features = true }
 
 # External Crates
 async-trait = { workspace = true }
 futures = { workspace = true }
-hex-literal = { workspace = true, default-features = true }
 is_executable = { workspace = true }
 gum = { workspace = true, default-features = true }
 log = { workspace = true, default-features = true }
-schnellru = { workspace = true }
 serde = { features = ["derive"], workspace = true, default-features = true }
 serde_json = { workspace = true, default-features = true }
 thiserror = { workspace = true }
@@ -93,13 +83,11 @@ kvdb-rocksdb = { optional = true, workspace = true }
 parity-db = { optional = true, workspace = true }
 codec = { workspace = true, default-features = true }
 parking_lot = { workspace = true, default-features = true }
-bitvec = { optional = true, workspace = true, default-features = true }
 
 # Polkadot
 polkadot-core-primitives = { workspace = true, default-features = true }
 polkadot-node-core-parachains-inherent = { workspace = true, default-features = true }
 polkadot-overseer = { workspace = true, default-features = true }
-polkadot-parachain-primitives = { workspace = true, default-features = true }
 polkadot-primitives = { workspace = true, default-features = true }
 polkadot-node-primitives = { workspace = true, default-features = true }
 polkadot-rpc = { workspace = true, default-features = true }
@@ -189,13 +177,11 @@ full-node = [
 
 # Configure the native runtimes to use.
 westend-native = [
-	"bitvec",
 	"frame-metadata-hash-extension",
 	"westend-runtime",
 	"westend-runtime-constants",
 ]
 rococo-native = [
-	"bitvec",
 	"frame-metadata-hash-extension",
 	"rococo-runtime",
 	"rococo-runtime-constants",
@@ -211,26 +197,18 @@ metadata-hash = [
 runtime-benchmarks = [
 	"frame-benchmarking-cli/runtime-benchmarks",
 	"frame-benchmarking/runtime-benchmarks",
-	"frame-support/runtime-benchmarks",
 	"frame-system/runtime-benchmarks",
-	"pallet-babe/runtime-benchmarks",
-	"pallet-staking/runtime-benchmarks",
-	"polkadot-parachain-primitives/runtime-benchmarks",
 	"polkadot-primitives/runtime-benchmarks",
 	"polkadot-runtime-parachains/runtime-benchmarks",
 	"polkadot-test-client/runtime-benchmarks",
 	"rococo-runtime?/runtime-benchmarks",
-	"sc-client-db/runtime-benchmarks",
 	"sc-service/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
 	"westend-runtime?/runtime-benchmarks",
 	"xcm-runtime-apis/runtime-benchmarks",
 ]
 try-runtime = [
-	"frame-support/try-runtime",
 	"frame-system/try-runtime",
-	"pallet-babe/try-runtime",
-	"pallet-staking/try-runtime",
 	"pallet-transaction-payment/try-runtime",
 	"polkadot-runtime-parachains/try-runtime",
 	"rococo-runtime?/try-runtime",
